refactor: centralize hybrid pool assembly for cache linkers
