+++ python/sglang/srt/mem_cache/hybrid_cache/hybrid_pool_assembler.py
@@ -2,7 +2,9 @@ from __future__ import annotations
 
 import logging
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Callable, Optional
+from typing import TYPE_CHECKING, Any, Callable, NamedTuple, Optional
+
+import torch
 
 from sglang.srt.mem_cache.hicache_storage import (
     PoolHitPolicy,
@@ -28,6 +30,10 @@ from sglang.srt.mem_cache.pool_host.mha import (
 )
 from sglang.srt.mem_cache.pool_host.mla import MLATokenToKVPoolHost
 from sglang.srt.mem_cache.unified_cache.component_type import ComponentType
+from sglang.srt.mem_cache.unified_cache_linker import (
+    DevicePoolEntry,
+    DevicePoolGroup,
+)
 from sglang.srt.runtime_context import get_parallel
 
 if TYPE_CHECKING:
@@ -68,6 +74,179 @@ def _with_mtp_layer_mapping(
         transfer_layer_start + depth: target_device_layer_num + depth
         for depth in range(draft_layer_num)
     }
+
+
+class _DeepSeekV4LayerMappings(NamedTuple):
+    transfer_layer_num: int
+    full: dict[int, int]
+    swa: dict[int, int]
+    c4: dict[int, int]
+    c128: dict[int, int]
+    c4_state: dict[int, int]
+    c4_state_global_layers: list[int]
+
+
+def _resolve_deepseek_v4_layer_mappings(
+    kvcache: Any,
+) -> _DeepSeekV4LayerMappings:
+    transfer_layer_num = kvcache.end_layer - kvcache.start_layer
+    full = {layer: layer for layer in range(transfer_layer_num)}
+    swa = {} if getattr(kvcache, "_unified_kv", False) else full.copy()
+
+    c4, c128, c4_state_global_layers = {}, {}, []
+    for local_layer, item in enumerate(
+        kvcache.layer_mapping[kvcache.start_layer : kvcache.end_layer]
+    ):
+        if item.compress_ratio == 4:
+            c4[local_layer] = item.compress_layer_id
+            c4_state_global_layers.append(kvcache.start_layer + local_layer)
+        elif item.compress_ratio == 128:
+            c128[local_layer] = item.compress_layer_id
+
+    return _DeepSeekV4LayerMappings(
+        transfer_layer_num=transfer_layer_num,
+        full=full,
+        swa=swa,
+        c4=c4,
+        c128=c128,
+        c4_state={layer: index for index, layer in enumerate(c4)},
+        c4_state_global_layers=c4_state_global_layers,
+    )
+
+
+def _deepseek_v4_state_views(state_pools: list[Any], global_layers: list[int]):
+    views = []
+    for layer in global_layers:
+        pool = state_pools[layer]
+        state = pool.kv_score_buffer.kv_score
+        ring = int(pool.ring_size)
+        usable = state.shape[0] // ring * ring
+        views.append(
+            state.view(torch.uint8)
+            .reshape(state.shape[0], -1)[:usable]
+            .reshape(usable // ring, -1)
+        )
+    return views
+
+
+def _build_deepseek_v4_device_pool_group(
+    kvcache: Any,
+    page_size: int,
+    mappings: _DeepSeekV4LayerMappings,
+) -> DevicePoolGroup:
+    from sglang.srt.mem_cache.deepseek_v4_memory_pool import HiSparseC4DevicePool
+
+    if getattr(kvcache, "_unified_kv", False) or isinstance(
+        kvcache.c4_kv_pool, HiSparseC4DevicePool
+    ):
+        raise ValueError(
+            "The direct external linker does not support unified-KV or HiSparse."
+        )
+    if kvcache.swa_page_size != page_size:
+        raise ValueError(
+            "DeepSeek V4 SWA page size must match the tree page size: "
+            f"{kvcache.swa_page_size} != {page_size}."
+        )
+
+    entries = [
+        DevicePoolEntry(
+            name=PoolName.SWA,
+            indices_from_pool=PoolName.SWA,
+            device_pool=kvcache.swa_kv_pool,
+            components=[kvcache.swa_kv_pool.kv_buffer],
+            layer_mapping=mappings.swa,
+            page_size=page_size,
+            rows_are_pages=True,
+        )
+    ]
+
+    def add(name, source, pool, buffers, layer_mapping):
+        if layer_mapping:
+            entries.append(
+                DevicePoolEntry(
+                    name=name,
+                    indices_from_pool=source,
+                    device_pool=pool,
+                    components=[buffers],
+                    layer_mapping=layer_mapping,
+                    page_size=page_size,
+                    rows_are_pages=True,
+                )
+            )
+
+    add(
+        PoolName.DEEPSEEK_V4_C4,
+        PoolName.KV,
+        kvcache.c4_kv_pool,
+        kvcache.c4_kv_pool.kv_buffer,
+        mappings.c4,
+    )
+    add(
+        PoolName.DEEPSEEK_V4_C4_INDEXER,
+        PoolName.KV,
+        kvcache.c4_indexer_kv_pool,
+        kvcache.c4_indexer_kv_pool.index_k_with_scale_buffer,
+        mappings.c4,
+    )
+    add(
+        PoolName.DEEPSEEK_V4_C128,
+        PoolName.KV,
+        kvcache.c128_kv_pool,
+        kvcache.c128_kv_pool.kv_buffer,
+        mappings.c128,
+    )
+    add(
+        PoolName.DEEPSEEK_V4_C4_STATE,
+        PoolName.SWA,
+        kvcache.compress_state_pools,
+        _deepseek_v4_state_views(
+            kvcache.compress_state_pools,
+            mappings.c4_state_global_layers,
+        ),
+        mappings.c4_state,
+    )
+    add(
+        PoolName.DEEPSEEK_V4_C4_INDEXER_STATE,
+        PoolName.SWA,
+        kvcache.indexer_compress_state_pools,
+        _deepseek_v4_state_views(
+            kvcache.indexer_compress_state_pools,
+            mappings.c4_state_global_layers,
+        ),
+        mappings.c4_state,
+    )
+    return DevicePoolGroup(entries, mappings.transfer_layer_num, page_size)
+
+
+def _build_dsa_device_pool_group(kvcache: Any, page_size: int) -> DevicePoolGroup:
+    if kvcache.page_size != page_size:
+        raise ValueError(
+            "DSA KV page size must match the tree page size: "
+            f"{kvcache.page_size} != {page_size}."
+        )
+    num_layers = kvcache.layer_num
+    identity = {layer: layer for layer in range(num_layers)}
+    entries = [
+        DevicePoolEntry(
+            name=PoolName.KV,
+            indices_from_pool=PoolName.KV,
+            device_pool=kvcache,
+            components=[kvcache.kv_buffer],
+            layer_mapping=identity,
+            page_size=page_size,
+            rows_are_pages=False,
+        ),
+        DevicePoolEntry(
+            name=PoolName.INDEXER,
+            indices_from_pool=PoolName.KV,
+            device_pool=kvcache,
+            components=[kvcache.index_k_with_scale_buffer],
+            layer_mapping=identity,
+            page_size=page_size,
+            rows_are_pages=True,
+        ),
+    ]
+    return DevicePoolGroup(entries, num_layers, page_size)
 
 
 def build_kv_host_pool(
@@ -370,10 +549,12 @@ def build_deepseek_v4_hicache_stack(
     model_name: Optional[str] = None,
     storage_backend_extra_config: Optional[dict] = None,
     enable_storage_metrics: bool = False,
+    layer_mappings: Optional[_DeepSeekV4LayerMappings] = None,
 ) -> tuple[HostPoolGroup, HybridCacheController]:
     page_size = params.page_size
-    transfer_layer_num = kvcache.end_layer - kvcache.start_layer
-    full_layer_mapping = {layer_id: layer_id for layer_id in range(transfer_layer_num)}
+    layer_mappings = layer_mappings or _resolve_deepseek_v4_layer_mappings(kvcache)
+    transfer_layer_num = layer_mappings.transfer_layer_num
+    full_layer_mapping = layer_mappings.full
 
     is_unified_kv = getattr(kvcache, "_unified_kv", False)
     mtp_swa_device_buffers = []
@@ -388,9 +569,7 @@ def build_deepseek_v4_hicache_stack(
                 f"got {len(kvcache.swa_kv_pool.kv_buffer)} buffers for "
                 f"{transfer_layer_num} local layers"
             )
-        swa_layer_mapping = {
-            layer_id: layer_id for layer_id in range(transfer_layer_num)
-        }
+        swa_layer_mapping = layer_mappings.swa
         # Keep every uncompressed draft SWA layer after the target SWA layers.
         # NextN has one layer per pool, while DSpark keeps all stages in one pool.
         mtp_swa_device_buffers = [
@@ -405,24 +584,10 @@ def build_deepseek_v4_hicache_stack(
             draft_layer_num=len(mtp_swa_device_buffers),
         )
 
-    c4_layer_mapping = {}
-    c128_layer_mapping = {}
-    c4_state_local_layers = []
-    c4_state_global_layers = []
-    for local_layer_id, layer_item in enumerate(
-        kvcache.layer_mapping[kvcache.start_layer : kvcache.end_layer]
-    ):
-        global_layer_id = kvcache.start_layer + local_layer_id
-        if layer_item.compress_ratio == 4:
-            c4_layer_mapping[local_layer_id] = layer_item.compress_layer_id
-            c4_state_local_layers.append(local_layer_id)
-            c4_state_global_layers.append(global_layer_id)
-        elif layer_item.compress_ratio == 128:
-            c128_layer_mapping[local_layer_id] = layer_item.compress_layer_id
-
-    c4_state_mapping = {
-        layer_id: local_id for local_id, layer_id in enumerate(c4_state_local_layers)
-    }
+    c4_layer_mapping = layer_mappings.c4
+    c128_layer_mapping = layer_mappings.c128
+    c4_state_mapping = layer_mappings.c4_state
+    c4_state_global_layers = layer_mappings.c4_state_global_layers
     num_host_pages, swa_num_host_pages = _deepseek_v4_num_host_pages(
         params=params,
         server_args=server_args,
@@ -1092,6 +1257,18 @@ class StackStrategy:
     def matches(self, kvcache: Any, components: set[ComponentType]) -> bool:
         raise NotImplementedError
 
+    def build_direct_linker_pool_group(
+        self,
+        *,
+        kvcache: Any,
+        params: CacheInitParams,
+        page_size: int,
+    ):
+        raise ValueError(
+            "The selected hybrid pool strategy does not support the direct "
+            f"external linker: {type(self).__name__}."
+        )
+
     def build(
         self,
         *,
@@ -1120,6 +1297,13 @@ class _DeepSeekV4Strategy(StackStrategy):
             ComponentType.SWA,
         }
 
+    def build_direct_linker_pool_group(self, *, kvcache, params, page_size):
+        return _build_deepseek_v4_device_pool_group(
+            kvcache,
+            page_size,
+            _resolve_deepseek_v4_layer_mappings(kvcache),
+        )
+
     def build(
         self,
         *,
@@ -1136,6 +1320,7 @@ class _DeepSeekV4Strategy(StackStrategy):
     ):
         from sglang.srt.mem_cache.base_prefix_cache import EvictParams
 
+        layer_mappings = _resolve_deepseek_v4_layer_mappings(kvcache)
         host_pool_group, cache_controller = build_deepseek_v4_hicache_stack(
             params=params,
             server_args=server_args,
@@ -1148,18 +1333,19 @@ class _DeepSeekV4Strategy(StackStrategy):
             model_name=model_name,
             storage_backend_extra_config=storage_backend_extra_config,
             enable_storage_metrics=enable_storage_metrics,
+            layer_mappings=layer_mappings,
         )
         sidecars = [
             SidecarPoolSpec(
                 pool_name=name,
-                indices_from_pool=src,
+                indices_from_pool=source,
                 hit_policy=(
                     PoolHitPolicy.TRAILING_PAGES
-                    if src == PoolName.SWA
+                    if source == PoolName.SWA
                     else PoolHitPolicy.ALL_PAGES
                 ),
             )
-            for name, src in (
+            for name, source in (
                 (PoolName.DEEPSEEK_V4_C4, PoolName.KV),
                 (PoolName.DEEPSEEK_V4_C4_INDEXER, PoolName.KV),
                 (PoolName.DEEPSEEK_V4_C128, PoolName.KV),
@@ -1394,6 +1580,9 @@ class _DsaStrategy(StackStrategy):
             ComponentType.FULL
         }
 
+    def build_direct_linker_pool_group(self, *, kvcache, params, page_size):
+        return _build_dsa_device_pool_group(kvcache, page_size)
+
     def build(
         self,
         *,
@@ -1599,6 +1788,21 @@ def _select_strategy(kvcache: Any, components: set[ComponentType]) -> StackStrat
     raise AssertionError(
         f"No matching HiCache strategy for kvcache={type(kvcache).__name__}, "
         f"components={sorted(c.name for c in components)}"
+    )
+
+
+def resolve_hybrid_device_pool_group(
+    *,
+    kvcache: Any,
+    page_size: int,
+    params: CacheInitParams,
+    components: set[ComponentType],
+):
+    """Materialize a direct-linker pool group through the assembler registry."""
+    return _select_strategy(kvcache, components).build_direct_linker_pool_group(
+        kvcache=kvcache,
+        params=params,
+        page_size=page_size,
     )
